Move link delete code to `deleteLink()` helper
